MCP-GITHUB/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
fastmcp==0.1.11
uvicorn==0.29.0